Added sample data and MVVM in UI

+++ MEMS Analyzer/Content/Data/SensorData.cs
@@ -9,16 +9,16 @@ namespace MEMS_Analyzer.Content.Data
     public class SensorData
     {
         public int id { get; set; }
-        public float accelX { get; set; }
-        public float accelY { get; set; }
-        public float accelZ { get; set; }
-        public float gyroX { get; set; }
-        public float gyroY { get; set; }
-        public float gyroZ { get; set; }
-        public float magnetoX { get; set; }
-        public float magnetoY { get; set; }
-        public float magnetoZ { get; set; }
-        public float airPressure { get; set; }
-        public float airTemp { get; set; }
+        public double accelX { get; set; }
+        public double accelY { get; set; }
+        public double accelZ { get; set; }
+        public double gyroX { get; set; }
+        public double gyroY { get; set; }
+        public double gyroZ { get; set; }
+        public double magnetoX { get; set; }
+        public double magnetoY { get; set; }
+        public double magnetoZ { get; set; }
+        public double airPressure { get; set; }
+        public double airTemp { get; set; }
     }
 }
